+++ client/src/index.tsx
@@ -9,6 +9,8 @@ const iFrame = <iframe src='pgadmin.html'
                         frameBorder={0}
                         width={"100%"}
                         height={"870"}
+                        name='pgadmin'
+                        id='pgadmin'
                         sandbox='allow-forms allow-modals allow-orientation-lock allow-pointer-lock allow-popups allow-popups-to-escape-sandbox allow-presentation allow-same-origin allow-scripts allow-top-navigation allow-top-navigation-by-user-activation' />;
 ReactDOM.render(
   <React.StrictMode>