+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## v6.10.0
 
-- Redefine Home Page
+- Redefined Home Page
+- Added progress indicator to the Home Page
 - Upgrade to PGAdmin4 6.10 docker image with dark mode enabled by default
 - More screenshots
 - Desktop Extension full description
+- Included Documentation in the Desktop Extension
 
 ## v6.9.0